+++ DigiKala.DataAccessLayer/Entities/Banner.cs
@@ -28,7 +28,7 @@ namespace DigiKala.DataAccessLayer.Entities
         public string Desc { get; set; }
 
         [Display(Name = "مبلغ اجاره")]
-        public int Price { get; set; }
+        public ulong Price { get; set; }
 
         [Display(Name = "تعداد روز")]
         public int Day { get; set; }
